moving stuff into the planarity core package
